pyproject.toml
[tool.poetry]
name = "dcspray"
version = "1.0.0"
description = "DRACOON Sprayer - Branding utility for DRACOON"
authors = ["Octavio Simone <70800577+unbekanntes-pferd@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-resize-image = "^1.1.20"
dracoon = "^1.8.0"
Pillow = "^9.2.0"
pydantic = "^1.10.2"
typer = "^0.6.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dcspray = "dcspray.clip:app"
